fix(io): slightly optimize loading/storing i/o
Move the `if` check out of the main loop so we don't branch on each pass and
use bitwise operations to get the 8-byte chunk count & remainder. The resulting
generated code is a little bit smaller (by 2k bytes or so) but runs about the
same speed. (This is unsurprising in retrospect: these are the sorts of changes
an optimizing compiler might perform!) I used hyperfine [1] and the extism CLI
to test `count_vowels` 100 times against `/usr/share/dict/words`.

This was useful primarily from a "getting my fingerprints on the problem space"
perspective.

Additionally, fix clippy lints.

[1]: https://github.com/sharkdp/hyperfine